+++ packages/ibm-products-web-components/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.12.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-web-components@0.12.0-rc.0...@carbon/ibm-products-web-components@0.12.0) (2025-04-02)
+
+**Note:** Version bump only for package @carbon/ibm-products-web-components
+
+
+
+
+
+# [0.12.0-rc.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-web-components@0.11.0-rc.0...@carbon/ibm-products-web-components@0.12.0-rc.0) (2025-03-24)
+
+
+### Bug Fixes
+
+* **useravatar:** renderIcon prop changed to slot ([#7099](https://github.com/carbon-design-system/ibm-products/issues/7099)) ([2745f90](https://github.com/carbon-design-system/ibm-products/commit/2745f905384ab446cccd0a437c41fe1e040ec6d8))
+
+
+### Features
+
+* **custom-element:** new es build module ([#7090](https://github.com/carbon-design-system/ibm-products/issues/7090)) ([ab5562b](https://github.com/carbon-design-system/ibm-products/commit/ab5562b03342bbf23d679cff9c1b1ab79e9c9c39))
+* **web-components:** adds FullPageError to web components ([#6910](https://github.com/carbon-design-system/ibm-products/issues/6910)) ([55a3a2f](https://github.com/carbon-design-system/ibm-products/commit/55a3a2f94757610ae5f69fb4e003ebbcbb78f69b))
+
+
+### Reverts
+
+* Revert "feat(web-components): adds FullPageError to web components (#6910)" (#7153) ([1130cfa](https://github.com/carbon-design-system/ibm-products/commit/1130cfa404f42a8c2bf094caa571fde27c7e4ee3)), closes [#6910](https://github.com/carbon-design-system/ibm-products/issues/6910) [#7153](https://github.com/carbon-design-system/ibm-products/issues/7153)
+
+
+
+
+
+# [0.11.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-web-components@0.11.0-rc.1...@carbon/ibm-products-web-components@0.11.0) (2025-03-19)
+
+**Note:** Version bump only for package @carbon/ibm-products-web-components
+
+
+
+
+
+# [0.11.0-rc.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-web-components@0.11.0-rc.0...@carbon/ibm-products-web-components@0.11.0-rc.1) (2025-03-12)
+
+
+### Features
+
+* **custom-element:** new es build module ([#7090](https://github.com/carbon-design-system/ibm-products/issues/7090)) ([ab5562b](https://github.com/carbon-design-system/ibm-products/commit/ab5562b03342bbf23d679cff9c1b1ab79e9c9c39))
+
+
+
+
+
 # [0.11.0-rc.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-web-components@0.3.0-rc.0...@carbon/ibm-products-web-components@0.11.0-rc.0) (2025-03-10)